+++ vendor-schemas.py
@@ -1,4 +1,5 @@
 #!/usr/bin/env python3
+import datetime
 import hashlib
 import shlex
 
@@ -6,6 +7,7 @@ import requests
 import ruamel.yaml
 
 yaml = ruamel.yaml.YAML(typ="safe")
+today = datetime.datetime.today().strftime("%Y-%m-%d")
 
 
 def entry2schema(entry):
@@ -31,10 +33,20 @@ def iter_schema_urls():
             yield (hook_id, schema)
 
 
-def download_schema(hookname, schema_url):
+def download_schema(hookname, schema_url) -> bool:
     res = requests.get(schema_url)
     sha = hashlib.sha256()
     sha.update(res.content)
+    new_digest = sha.hexdigest()
+
+    with open(
+        f"src/check_jsonschema/builtin_schemas/vendor/{hookname}.sha256", "r"
+    ) as fp:
+        prev_digest = fp.read().strip()
+
+    if new_digest == prev_digest:
+        return False
+
     with open(
         f"src/check_jsonschema/builtin_schemas/vendor/{hookname}.json", "wb"
     ) as fp:
@@ -42,15 +54,36 @@ def download_schema(hookname, schema_url):
     with open(
         f"src/check_jsonschema/builtin_schemas/vendor/{hookname}.sha256", "w"
     ) as fp:
-        fp.write(sha.hexdigest())
+        fp.write(new_digest)
+
+    return True
+
+
+def update_changelog():
+    with open("CHANGELOG.md", encoding="utf-8") as fp:
+        content = fp.read()
+    content = content.replace(
+        """
+<!-- vendor-insert-here -->
+""",
+        f"""
+<!-- vendor-insert-here -->
+- update vendored schemas ({today})
+""",
+    )
+    with open("CHANGELOG.md", "w", encoding="utf-8") as fp:
+        fp.write(content)
 
 
 def main():
+    made_changes = False
     for hook_id, schema_url in iter_schema_urls():
         hookname = hook_id
         if hookname.startswith("check-"):
             hookname = hookname[6:].replace("_", "-")
-        download_schema(hookname, schema_url)
+        made_changes = made_changes or download_schema(hookname, schema_url)
+    if made_changes:
+        update_changelog()
 
 
 if __name__ == "__main__":